+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Create `Catalog` CRs for the integration test.
+
 ## [0.10.3] - 2021-03-09
 
 ### Fixed